Transport S4b (HTTP side): session artifacts/deletes/worktrees onto the neutral api core
Second S4 slice, HTTP commit (design \$6 S4):

- Neutral cores in routes_sessions.rs: session_delete_api_response (the
  five wire shapes' shared content; the delete tail keeps its historical
  wildcard-CORS-before-Cache-Control order),
  session_report_api_response (Bytes lane: the zip under its attachment
  tail, with the tunnel's byte_stream_end.result sidecar as meta) plus
  session_text_plain_response for the artifact error wordings,
  session_recordings_api_response, and
  session_recording_listing_asset_api_response (the tunnel twin's
  segments/playlist.m3u8 asset vocabulary through the shared resolver);
  worktrees_{list,inspect,scan,remove}_api_response own the inventory
  bodies plus the shared cache side-effects.
- The recording/frame asset content core — RecordingAsset, its
  resolvers, safety predicates, and range readers — moves verbatim from
  dashboard_control/api_media.rs into
  web_gateway/session_catalog/external_rows.rs with a pub(crate) use at
  the old location; segment/frame FILE leaves stay on their
  transport-owned carriage (HTTP full read vs the tunnel's ranged,
  capped byte streams).
- handle_session_delete and the worktrees quartet become shims taking
  the row's CORS posture from dispatch; the sub-router's report,
  recordings-list, and segments/playlist arms collapse onto the cores.
  The report branch now consumes session_report_zip_for_request (the
  shared core the tunnel already used) with per-lane error framing kept
  byte-exact (wildcard json invalid-id; text/plain not-found/build).

All 45 golden transcripts (S4a's fifteen + this slice's ten sets) pass
unchanged — byte-identical conversion.
